requirements.txt
# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Elasticsearch
elasticsearch>=8.11.0,<9.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client for Ollama
httpx>=0.25.0
aiohttp>=3.9.0

# Text processing
tiktoken>=0.5.2
langdetect>=1.0.9

# PDF processing with PaddleOCR
unstructured[pdf,ocr-paddle]>=0.11.0

# MinIO storage
minio>=7.2.0

# Configuration
python-dotenv>=1.0.0

# Utilities
uuid6>=2024.1.12

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
